New proposal for multi-step control flow alignment
This is 100% open to discussion.  In brace languages, I prefer to see subsequent keywords like `else` aligned with the original `if`.  For me it's easier to see what's happening when lines begin with keywords instead of braces.  Thoughts?

PS I noticed the original style included empty lines before the `else`, like so:

```
if (condition) {
  doSomething();

} else {
  doSomethingElse();
}
```

I think my proposal retains the spirit of the newline.  I have a line that's empty except for a single brace, which provides similar visual space to the original.